This commit is a fix of passing projection arguments. Initially I didn't have the option to use multiple projections. Upon paper review I added the arguments in sucha  wa y that every funciton enumerated all teh various argumetn (lot0, lat1, lon0,n, r) even though the vast majority of use cases won't use these argumetns. upon code review in BioConductor it was sugessted that all of these argumetns be removed i nfavor of a genral argument. This is finally finxin that through the use of the special "..." variable.
If a user tries a proecjtion that requires a special varialbe they will be prompted for this variabel.

In putting together this code change, the newyorker projection if borken. For now I am simple removing if from teh lsit of supported projections although I will consider restoring it at another time.